# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PufferPanel daemon service",
        "title": "Pufferd API",
        "contact": {
            "name": "PufferPanel",
            "url": "https://pufferpanel.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
        "/server/{id}": {
            "post": {
                "description": "Starts the given server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Starts server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server Identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apufferi.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/apufferi.Error"
                }
            }
        }
    }
}